+++ .env.example
@@ -1,2 +1,3 @@
 MONGO_URI=mongodb://localhost:27017
 JWT_SECRET_KEY=JWT_SECRET_KEY
+PAYMENT_ACCOUNT_NUMBER=TNBC_ACCOUNT_NUMBER
